wefix/ifix/wefix/requirements.txt
Django
psycopg2
psycopg2-binary
djangorestframework
django-cors-headers
pika
python-dotenv